support for 64 bit rates 64 bit platforms
Without 64 bit numbers, we cannot specify link capacities greater than
34 gbps, which prevents click usage for our use case. This commit bumps
the bigint::limb_type from uint32_t to unsigned long long. This should
enable use on both 32 and 64 bit architectures.

Light testing with BandwidthRatedUnqueue shows that exceed the 32 bit
threshold impose the expected rates.

+++ elements/standard/ratedsource.cc
@@ -50,7 +50,7 @@ RatedSource::configure(Vector<String> &conf, ErrorHandler *errh)
     String data =
       "Random bullshit in a packet, at least 64 bytes long. Well, now it is.";
     unsigned rate = 10;
-    unsigned bandwidth = 0;
+    unsigned long long bandwidth = 0;
     int limit = -1;
     int datasize = -1;
     bool active = true, stop = false;
